Updates for sa isolate run

+++ workflow/scripts/04_virus_identification_analysis.py
@@ -12,7 +12,7 @@ genomad_report = report[(report['virus_score'] >= snakemake.params.genomad_score
 tools.append('geNomad')
 tool_counts[['assembly', 'geNomad']] = genomad_report.groupby(['assembly'], as_index=False).count()[['assembly', 'vls_id']]
 
-external_report = report[(report['identity'] >= snakemake.params.min_mash_score) & (report['shared-hashes'] >= snakemake.params.min_mash_hashes) & (report['median-multiplicity'] >= snakemake.params.min_mash_multiplicity)]
+external_report = report[(report['identity'].astype(float) >= snakemake.params.min_mash_score) & (report['shared-hashes'].astype(float) >= snakemake.params.min_mash_hashes) & (report['median-multiplicity'].astype(float) >= snakemake.params.min_mash_multiplicity)]
 external_report.rename(columns={'shared-hashes':'shared_hashes', 'median-multiplicity':'median_multiplicity'}, inplace=True)
 tools.append('external')
 tool_counts[['assembly','external']] = external_report.groupby(['assembly'], as_index=False).count()[['assembly', 'vls_id']]
